% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFit.R
\name{delayed_initialization}
\alias{delayed_initialization}
\title{Initialize the parameters of a Negative Binomial regression model with a 
DelayedArray object}
\usage{
delayed_initialization(cluster, children, model, verbose, Y)
}
\arguments{
\item{cluster}{The PSOCK cluster}

\item{children}{Number of child process.}

\item{model}{The newmodel object}

\item{verbose}{Print proc time}

\item{Y}{Is the data matrix}
}
\value{
It does not return anything, the parameters
 are update internally as these are shared object
}
\description{
It initialize gamma and beta using a Ridge Regression and W and alpha using PCA
}
\keyword{internal}
