% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getBeta}
\alias{getBeta}
\title{Returns the matrix of paramters beta}
\usage{
getBeta(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the matrix of beta parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of parameters associated with X
}
\examples{
a <- nbModel(n=5, J=10)
getBeta(a)
}
