% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getEpsilon_alpha}
\alias{getEpsilon_alpha}
\title{Returns the vector of regularization parameter for alpha}
\usage{
getEpsilon_alpha(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the regularization parameters for \code{alpha}.
}
\description{
Given an object describing a nb model, returns a vector of size the number
of rows in the parameter \code{alpha} with the regularization parameters
associated to each row.
}
\examples{
a <- nbModel(n=5, J=10)
getEpsilon_alpha(a)
}
