% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getV}
\alias{getV}
\title{Returns the gene-level design matrix for mu}
\usage{
getV(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the gene-level design matrix for mu
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the gene-level design matrix for mu
}
\examples{
a <- nbModel(n=5, J=10)
getV(a)
}
