% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getX}
\alias{getX}
\title{Returns the sample-level design matrix for mu}
\usage{
getX(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}

\item{...}{Additional parameters.}
}
\value{
the sample-level design matrix for mu
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the sample-level design matrix for mu
}
\examples{
a <- nbModel(n=5, J=10)
getX(a)
}
