% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFit.R
\name{initialization}
\alias{initialization}
\title{Initialize the parameters of a Negative Binomial regression model}
\usage{
initialization(cluster, children, model, verbose, Y)
}
\arguments{
\item{cluster}{The PSOCK cluster}

\item{children}{Number of child process.}

\item{model}{The newmodel object}

\item{verbose}{Print proc time}
}
\description{
It initialize gamma and beta using a Ridge Regression and W and alpha using PCA
}
\keyword{internal}
