% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newWave.R
\name{nb.loglik.matrix}
\alias{nb.loglik.matrix}
\title{Log-likelihood of the zero-inflated negative binomial model for each entry
in the matrix of counts}
\usage{
nb.loglik.matrix(model, x)
}
\arguments{
\item{model}{the newmodel}

\item{x}{the matrix of counts}
}
\value{
the matrix of log-likelihood of the model.
}
\description{
Given a matrix of counts, this function computes the
log-probabilities of the counts under a zero-inflated negative binomial
(NB) model. For each count, the NB distribution is parametrized by three
parameters: the mean value and the dispersion of the negative binomial
distribution, and the probability of the zero component.
}
