% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zinb_fit.R
\name{nbFit}
\alias{nbFit}
\alias{nbFit,SummarizedExperiment-method}
\alias{nbFit,matrix-method}
\alias{nbFit,dgCMatrix-method}
\title{Fit a nb regression model}
\usage{
nbFit(Y, ...)

\S4method{nbFit}{SummarizedExperiment}(
  Y,
  X,
  V,
  K = 2,
  which_assay,
  commondispersion = T,
  verbose = FALSE,
  maxiter_optimize = 100,
  stop_epsilon = 1e-04,
  children = 1,
  random_init = FALSE,
  random_start = FALSE,
  n_gene_disp = NULL,
  n_cell_par = NULL,
  n_gene_par = NULL,
  cross_batch = F,
  multi_obs = 1,
  ...
)

\S4method{nbFit}{matrix}(
  Y,
  X,
  V,
  K,
  commondispersion,
  verbose,
  maxiter_optimize,
  stop_epsilon,
  children,
  random_init,
  random_start,
  n_gene_disp,
  n_cell_par,
  n_gene_par,
  cross_batch,
  multi_obs,
  ...
)

\S4method{nbFit}{dgCMatrix}(Y, ...)
}
\arguments{
\item{Y}{The data (genes in rows, samples in columns).}

\item{...}{Additional parameters to describe the model, see
\code{\link{nbModel}}.}

\item{X}{The design matrix containing sample-level covariates, one sample per
row. If missing, X will contain only an intercept. If Y is a
SummarizedExperiment object, X can be a formula using the variables in the
colData slot of Y.}

\item{V}{The design matrix containing gene-level covariates, one gene
per row. If missing, V will contain only an intercept. If Y is a
SummarizedExperiment object, V can be a formula using the variables in the
rowData slot of Y.}

\item{K}{integer. Number of latent factors, by default equal to two.}

\item{which_assay}{numeric or character. Which assay of Y to use (only if Y
is a SummarizedExperiment).}

\item{commondispersion}{logical. If FALSE genewise dispersion parameter is 
estimated.}

\item{verbose}{Print helpful messages.}

\item{maxiter_optimize}{maximum number of iterations for the optimization
step (default 100).}

\item{stop_epsilon}{stopping criterion in the optimization step,
when the relative gain in likelihood is below epsilon (default 0.0001).}

\item{children}{Number of threads to use for computation.}

\item{random_init}{logical, if TRUE no initialization component is performed.}

\item{random_start}{logical, if TRUE starting values for parameter are sampled
by a normal distribuitions.}

\item{n_gene_disp}{gene batch dimension in dispersion parameter optimization,
by default all genes are used.}

\item{n_cell_par}{cell batch dimension in the left part optimization, 
by default all cell are used.}

\item{n_gene_par}{gene batch dimension in the right part optimization, 
by default all genes are used.}

\item{cross_batch}{logical, if TRUE during the right part optimization 
only a subset of cell are used for optimization and in the left optimization
only a subset of genes.}

\item{multi_obs}{numeric, number of parameters estimated simoultaneously.}
}
\value{
An object of class \code{nbModel} that has been fitted by penalized
  maximum likelihood on the data.
}
\description{
Given an object with the data, it fits a nb model.
}
\details{
By default, i.e., if no arguments other than \code{Y} are passed,
  the model is fitted with an intercept for the regression across-samples and
  one intercept for the regression across genes.

If Y is a Summarized experiment, the function uses the assay named
  "counts", if any, or the first assay.

Currently, if Y is a sparseMatrix, this calls the nbFit method on
  as.matrix(Y)
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: Y is a
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.

\item \code{matrix}: Y is a matrix of counts (genes in rows).

\item \code{dgCMatrix}: Y is a sparse matrix of counts (genes in rows).
}}

\examples{
se <- SummarizedExperiment(matrix(rpois(60, lambda=5), nrow=10, ncol=6),
                           colData = data.frame(bio = gl(2, 3)))

m <- nbFit(se, X=model.matrix(~bio, data=colData(se)))
bio <- gl(2, 3)
m <- nbFit(matrix(rpois(60, lambda=5), nrow=10, ncol=6),
             X=model.matrix(~bio))
}
\seealso{
\code{\link[stats]{model.matrix}}.
}
