% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/helper_methods.R
\docType{class}
\name{nbModel-class}
\alias{nbModel-class}
\alias{nbModel}
\alias{show,nbModel-method}
\alias{nSamples,nbModel-method}
\alias{nFeatures,nbModel-method}
\alias{nFactors,nbModel-method}
\alias{getX,nbModel-method}
\alias{getV,nbModel-method}
\alias{getLogMu,nbModel-method}
\alias{getMu,nbModel-method}
\alias{getZeta,nbModel-method}
\alias{getPhi,nbModel-method}
\alias{getTheta,nbModel-method}
\alias{getEpsilon_beta,nbModel-method}
\alias{getEpsilon_gamma,nbModel-method}
\alias{getEpsilon_W,nbModel-method}
\alias{getEpsilon_alpha,nbModel-method}
\alias{getEpsilon_zeta,nbModel-method}
\alias{getW,nbModel-method}
\alias{getBeta,nbModel-method}
\alias{getGamma,nbModel-method}
\alias{getAlpha,nbModel-method}
\title{Class nbModel}
\usage{
\S4method{show}{nbModel}(object)

\S4method{nSamples}{nbModel}(x)

\S4method{nFeatures}{nbModel}(x)

\S4method{nFactors}{nbModel}(x)

\S4method{getX}{nbModel}(object)

\S4method{getV}{nbModel}(object)

\S4method{getLogMu}{nbModel}(object)

\S4method{getMu}{nbModel}(object)

\S4method{getZeta}{nbModel}(object)

\S4method{getPhi}{nbModel}(object)

\S4method{getTheta}{nbModel}(object)

\S4method{getEpsilon_beta}{nbModel}(object)

\S4method{getEpsilon_gamma}{nbModel}(object)

\S4method{getEpsilon_W}{nbModel}(object)

\S4method{getEpsilon_alpha}{nbModel}(object)

\S4method{getEpsilon_zeta}{nbModel}(object)

\S4method{getW}{nbModel}(object)

\S4method{getBeta}{nbModel}(object)

\S4method{getGamma}{nbModel}(object)

\S4method{getAlpha}{nbModel}(object)
}
\arguments{
\item{object}{an object of class \code{nbModel}.}

\item{x}{an object of class \code{nbModel}.}
}
\value{
\code{nSamples} returns the number of samples; \code{nFeatures}
returns the number of features; \code{nFactors} returns the number of latent
factors.
}
\description{
Objects of this class store all the values needed to work with a
negative binomial model, as described in the vignette.
They contain all information to fit a model by penalized maximum likelihood
or simulate data from a model.
}
\details{
For the full description of the model see the model vignette.
  Internally, the slots are checked so that the matrices are of the
  appropriate dimensions: in particular, \code{X}, \code{O}
  and \code{W} need to have \code{n} rows, \code{V} needs to have \code{J}
  rows, \code{zeta} must be of length \code{J}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show useful info on the object.

\item \code{nSamples}: returns the number of samples.

\item \code{nFeatures}: returns the number of features.

\item \code{nFactors}: returns the number of latent factors.

\item \code{getX}: returns the sample-level design matrix for mu.

\item \code{getV}: returns the gene-level design matrix for mu.

\item \code{getLogMu}: returns the logarithm of the mean of the non-zero
component.

\item \code{getMu}: returns the mean of the non-zero component.

\item \code{getZeta}: returns the log of the inverse of the dispersion
parameter.

\item \code{getPhi}: returns the dispersion parameter.

\item \code{getTheta}: returns the inverse of the dispersion parameter.

\item \code{getEpsilon_beta}: returns the regularization parameters for
\code{beta}.

\item \code{getEpsilon_gamma}: returns the regularization parameters for
\code{gamma}.

\item \code{getEpsilon_W}: returns the regularization parameters for
\code{W}.

\item \code{getEpsilon_alpha}: returns the regularization parameters for
\code{alpha}.

\item \code{getEpsilon_zeta}: returns the regularization parameters for
\code{zeta}.

\item \code{getW}: returns the matrix W of inferred sample-level
covariates.

\item \code{getBeta}: returns the matrix beta of inferred parameters.

\item \code{getGamma}: returns the matrix gamma of inferred parameters.

\item \code{getAlpha}: returns the matrix alpha of inferred parameters.
}}

\section{Slots}{

\describe{
\item{\code{X}}{matrix. The design matrix containing sample-level covariates, one
sample per row.}

\item{\code{V}}{matrix. The design matrix containing gene-level covariates, one gene
per row.}

\item{\code{O}}{matrix. The offset matrix.}

\item{\code{X_intercept}}{logical. TRUE if X contains an intercept.}

\item{\code{V_intercept}}{logical. TRUE if V contains an intercept.}

\item{\code{W}}{matrix. The factors of sample-level latent factors.}

\item{\code{beta}}{matrix or NULL. The coefficients of X in the regression.}

\item{\code{gamma}}{matrix or NULL. The coefficients of V in the regression.}

\item{\code{alpha}}{matrix. The weight of sample-level latent factors.}

\item{\code{zeta}}{numeric. A vector of log of inverse dispersion parameters.}

\item{\code{epsilon_beta}}{nonnegative scalar. Regularization parameter for
beta}

\item{\code{epsilon_gamma}}{nonnegative scalar. Regularization parameter for
gamma}

\item{\code{epsilon_W}}{nonnegative scalar. Regularization parameter for W}

\item{\code{epsilon_alpha}}{nonnegative scalar. Regularization parameter for alpha}

\item{\code{epsilon_zeta}}{nonnegative scalar. Regularization parameter for zeta}
}}

