% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newAlpha}
\alias{newAlpha}
\title{Returns the matrix of paramters alpha}
\usage{
newAlpha(object, ...)
}
\arguments{
\item{object}{an object that describes a matrix of negative binomial
distributions.}

\item{...}{Additional parameters.}
}
\value{
the matrix of alpha parameters
}
\description{
Given an object that describes a matrix of negative binomial distributions,
returns the matrix of parameters associated with W for the mean part (mu)
}
\examples{
a <- newmodel(n=5, J=10)
newAlpha(a)
}
