% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newEpsilon_beta}
\alias{newEpsilon_beta}
\title{Returns the vector of regularization parameter for beta}
\usage{
newEpsilon_beta(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the regularization parameters for \code{beta}.
}
\description{
Given an object describing a nb model, returns a vector of size the number
of rows in the parameter \code{beta} with the regularization parameters
associated to each row.
}
\examples{
a <- newmodel(n=5, J=10)
newEpsilon_beta(a)
}
