% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newEpsilon_gamma}
\alias{newEpsilon_gamma}
\title{Returns the vector of regularization parameter for gamma}
\usage{
newEpsilon_gamma(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the regularization parameters for \code{gamma}.
}
\description{
Given an object describing a nb model, returns a vector of size the number
of columns in the parameter \code{gamma} with the regularization
parameters associated to each row.
}
\examples{
a <- newmodel(n=5, J=10)
newEpsilon_gamma(a)
}
