% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newEpsilon_W}
\alias{newEpsilon_W}
\title{Returns the vector of regularization parameter for W}
\usage{
newEpsilon_W(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the regularization parameters for \code{W}.
}
\description{
Given an object describing a nb model, returns a vector of size the number
of columns in the parameter \code{W} with the regularization
parameters associated to each column.
}
\examples{
a <- newmodel(n=5, J=10)
newEpsilon_W(a)
}
