% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newEpsilon_zeta}
\alias{newEpsilon_zeta}
\title{Returns the regularization parameter for the dispersion parameter}
\usage{
newEpsilon_zeta(object)
}
\arguments{
\item{object}{an object that describes a matrix of negative binomial
distributions.}
}
\value{
the regularization parameters for \code{zeta}.
}
\description{
The regularization parameter penalizes the variance of zeta, the log of
the dispersion parameters across samples.
}
\examples{
a <- newmodel(n=5, J=10)
newEpsilon_zeta(a)
}
