% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/newFit.R
\name{newFit}
\alias{newFit}
\alias{newFit,SummarizedExperiment-method}
\alias{newFit,matrix-method}
\alias{newFit,DelayedMatrix-method}
\alias{newFit,dgCMatrix-method}
\title{Fit a nb regression model}
\usage{
newFit(Y, ...)

\S4method{newFit}{SummarizedExperiment}(
  Y,
  X,
  V,
  K = 2,
  which_assay,
  commondispersion = T,
  verbose = FALSE,
  maxiter_optimize = 100,
  stop_epsilon = 1e-04,
  children = 1,
  random_init = FALSE,
  random_start = FALSE,
  n_gene_disp = NULL,
  n_cell_par = NULL,
  n_gene_par = NULL,
  cross_batch = F,
  ...
)

\S4method{newFit}{matrix}(
  Y,
  X,
  V,
  K = 2,
  commondispersion = T,
  verbose = FALSE,
  maxiter_optimize = 100,
  stop_epsilon = 1e-04,
  children = 1,
  random_init = FALSE,
  random_start = FALSE,
  n_gene_disp = NULL,
  n_cell_par = NULL,
  n_gene_par = NULL,
  cross_batch = F,
  ...
)

\S4method{newFit}{DelayedMatrix}(
  Y,
  X,
  V,
  K,
  commondispersion,
  verbose,
  nb_repeat,
  maxiter_optimize,
  stop_epsilon,
  children,
  random_start,
  n_gene_disp,
  n_cell_par,
  n_gene_par,
  ...
)

\S4method{newFit}{dgCMatrix}(Y, ...)
}
\arguments{
\item{Y}{The data (genes in rows, samples in columns).}

\item{...}{Additional parameters to describe the model, see
\code{\link{newmodel}}.}

\item{X}{The design matrix containing sample-level covariates, one sample per
row. If missing, X will contain only an intercept. If Y is a
SummarizedExperiment object, X can be a formula using the variables in the
colData slot of Y.}

\item{V}{The design matrix containing gene-level covariates, one gene
per row. If missing, V will contain only an intercept. If Y is a
SummarizedExperiment object, V can be a formula using the variables in the
rowData slot of Y.}

\item{K}{integer. Number of latent factors..}

\item{which_assay}{numeric or character. Which assay of Y to use (only if Y
is a SummarizedExperiment).}

\item{verbose}{Print helpful messages.}

\item{nb.repeat.initialize}{Number of iterations for the initialization of
beta and gamma.}

\item{maxiter.optimize}{maximum number of iterations for the optimization
step (default 25).}

\item{stop.epsilon.optimize}{stopping criterion in the optimization step,
when the relative gain in likelihood is below epsilon (default 0.0001).}
}
\value{
An object of class \code{newmodel} that has been fitted by penalized
  maximum likelihood on the data.
}
\description{
Given an object with the data, it fits a nb model.
}
\details{
By default, i.e., if no arguments other than \code{Y} are passed,
  the model is fitted with an intercept for the regression across-samples and
  one intercept for the regression across genes.

If Y is a Summarized experiment, the function uses the assay named
  "counts", if any, or the first assay.

Currently, if Y is a sparseMatrix, this calls the newFit method on
  as.matrix(Y)
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: Y is a
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.

\item \code{matrix}: Y is a matrix of counts (genes in rows).

\item \code{DelayedMatrix}: Y is a DeleyedMatrix of counts (genes in rows).

\item \code{dgCMatrix}: Y is a sparse matrix of counts (genes in rows).
}}

\examples{
se <- SummarizedExperiment(matrix(rpois(60, lambda=5), nrow=10, ncol=6),
                           colData = data.frame(bio = gl(2, 3)))

m <- newFit(se, X=model.matrix(~bio, data=colData(se)))
bio <- gl(2, 3)
m <- newFit(matrix(rpois(60, lambda=5), nrow=10, ncol=6),
             X=model.matrix(~bio))
}
\seealso{
\code{\link[stats]{model.matrix}}.
}
