% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newMu}
\alias{newMu}
\title{Returns the matrix of mean parameters}
\usage{
newMu(object)
}
\arguments{
\item{object}{an object that describes a matrix of zero-inflated
distributions.}
}
\value{
the matrix of mean parameters
}
\description{
Given an object that describes a matrix of zero-inflated distributions,
returns the matrix of mean parameters.
}
\details{
Note that although the user interface of \code{\link{newFit}}
requires a J x n matrix, internally this is stored as a n x J matrix (i.e.,
samples in row and genes in column). Hence the parameter matrix returned by
this function is of n x J dimensions.
}
\examples{
a <- newmodel(n=5, J=10)
newMu(a)
}
