% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newTheta}
\alias{newTheta}
\title{Returns the vector of inverse dispersion parameters}
\usage{
newTheta(object)
}
\arguments{
\item{object}{an object that describes a matrix of negative binomial
distributions.}
}
\value{
the vector of inverse dispersion parameters theta
}
\description{
Given an object that describes a matrix of negative binomial negative binomial
distributions, returns the vector of inverse dispersion parameters
\code{theta}.
}
\examples{
a <- newmodel(n=5, J=10)
newTheta(a)
}
