% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newW}
\alias{newW}
\title{Returns the low-dimensional matrix of inferred sample-level covariates W}
\usage{
newW(object)
}
\arguments{
\item{object}{a \code{\linkS4class{newmodel}} object, typically the result
of \code{\link{newFit}}.}
}
\value{
the matrix \code{W} of inferred sample-level covariates.
}
\description{
Given an object that contains the fit of a nb-WaVE model, returns the
matrix \code{W} of low-dimensional matrix of inferred sample-level
covariates.
}
\examples{
a <- newmodel(n=5, J=10)
newW(a)
}
