% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newZeta}
\alias{newZeta}
\title{Returns the vector of log of inverse dispersion parameters}
\usage{
newZeta(object)
}
\arguments{
\item{object}{an object that describes a matrix of negative binomial
distributions.}
}
\value{
the vector \code{zeta} of log of inverse dispersion parameters
}
\description{
Given an object that describes a matrix of negative binomial negative binomial
distributions, returns the vector \code{zeta} of log of inverse dispersion
parameters
}
\examples{
a <- newmodel(n=5, J=10)
newZeta(a)
}
