% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{numberFactors}
\alias{numberFactors}
\title{Generic function that returns the number of latent factors}
\usage{
numberFactors(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model involving latent
factors}
}
\value{
the number of latent factors
}
\description{
Given an object that describes a dataset or a model involving latent factors,
this function returns the number of latent factors.
}
\examples{
a <- newmodel(n=5, J=10)
numberFactors(a)
}
