% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_fit.R
\name{optimization}
\alias{optimization}
\title{Optimize the parameters of a Negative Binomial regression model}
\usage{
optimization(
  cluster,
  children = 1,
  model,
  max_iter = 50,
  stop_epsilon = 1e-04,
  n_gene_disp = NULL,
  n_cell_par = NULL,
  n_gene_par = NULL,
  orthog = T,
  commondispersion = T,
  verbose,
  mode,
  cross_batch
)
}
\arguments{
\item{verbose}{print information (default FALSE)}

\item{m}{The model of class nbModel}

\item{Y}{The matrix of counts.}

\item{maxiter}{maximum number of iterations (default 25)}

\item{stop.epsilon}{stopping criterion, when the relative gain in
likelihood is below epsilon (default 0.0001)}
}
\value{
An object of class nbModel similar to the one given as argument
  with modified parameters alpha, beta, gamma, W.
}
\description{
The parameters of the model given as argument are optimized by penalized
maximum likelihood on the count matrix given as argument.
}
\examples{
Y = matrix(10, 3, 5)
m = nbModel(n=NROW(Y), J=NCOL(Y))
m = nbInitialize(m, Y)
m = nbOptimize(m, Y)
}
