% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFit.R
\name{setup}
\alias{setup}
\title{Setup the parameters of a Negative Binomial regression model}
\usage{
setup(cluster, model, random_start, children, random_init, verbose, Y, mode)
}
\arguments{
\item{random_start}{The type of start.}

\item{children}{Number of child process.}

\item{Y}{The matrix of counts.}

\item{m}{The model of class newmodel}
}
\description{
There are different type of starting values:
1.You can set all values to 0.
2.You can sample values from a gaussian distribution or a
chisq distibution for the dispersion parameters,
}
\details{
It creates different shared object and epxort them to the father
and the child process.
}
