% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{newPhi}
\alias{newPhi}
\title{Returns the vector of dispersion parameters}
\usage{
newPhi(object)
}
\arguments{
\item{object}{an object that describes a matrix of negative binomial.
distributions.}
}
\value{
the vector of dispersion parameters
}
\description{
Given an object that describes a matrix of negative binomial negative binomial
distributions, returns the vector of dispersion parameters \code{phi}.
}
\examples{
a <- newmodel(n=5, J=10)
newPhi(a)
}
