% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{numberSamples}
\alias{numberSamples}
\title{Generic function that returns the number of samples}
\usage{
numberSamples(x)
}
\arguments{
\item{x}{an object that describes a dataset or a model.}
}
\value{
the number of samples.
}
\description{
Given an object that describes a model or a dataset, it returns the number of
samples.
}
