% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFit.R
\name{optimization}
\alias{optimization}
\title{Optimize the parameters of a Negative Binomial regression model}
\usage{
optimization(
  cluster,
  children,
  model,
  max_iter,
  stop_epsilon,
  n_gene_disp,
  n_cell_par,
  n_gene_par,
  orthog,
  commondispersion,
  verbose,
  mode,
  cross_batch
)
}
\arguments{
\item{cluster}{The PSOCK cluster}

\item{children}{Number of child process}

\item{model}{newmodel item}

\item{max_iter}{maximum number of iterations}

\item{stop_epsilon}{stopping criterion, when the relative gain in
likelihood is below epsilon}

\item{verbose}{print information (default FALSE)}
}
\value{
An object of class newmodel similar to the one given as argument
  with modified parameters alpha, beta, gamma, W.
}
\description{
The parameters of the model given as argument are optimized by penalized
maximum likelihood on the count matrix given as argument.
}
