% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_fit.R
\name{setup}
\alias{setup}
\title{Setup the parameters of a Negative Binomial regression model}
\usage{
setup(
  cluster,
  model,
  random_start = F,
  children,
  random_init = F,
  verbose,
  Y,
  mode
)
}
\arguments{
\item{random_start}{The type of start.}

\item{children}{Number of child process.}

\item{Y}{The matrix of counts.}

\item{m}{The model of class nbModel}
}
\description{
There are different type of starting values:
1.You can set all values to 0.
2.You can sample values from a gaussian distribution or a
chisq distibution for the dispersion parameters,
}
\details{
It creates different shared object and exort them to the father
and the child process.
}
\examples{
Y <- matrix(rpois(60, lambda=2), 6, 10)
bio <- gl(2, 3)
time <- rnorm(6)
gc <- rnorm(10)
m <- nbModel(Y, X=model.matrix(~bio + time), V=model.matrix(~gc), K=1)
m <- nbInitialize(m, Y)

}
