% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFit.R
\name{setup}
\alias{setup}
\title{Setup the parameters of a Negative Binomial regression model
There are different type of starting values:
1.You can set all values to 0.
2.You can sample values from a gaussian distribution or a
chisq distibution for the dispersion parameters,}
\usage{
setup(cluster, model, random_start, children, random_init, verbose, Y_sh)
}
\arguments{
\item{cluster}{the PSOCK cluster object}

\item{model}{The model of class newmodel}

\item{random_start}{if TRUE the setup of parameters is a 
random samplig(default FALSE)}

\item{children}{Number of child process.}

\item{random_init}{if TRUE no initializations is done(default FALSE)}

\item{verbose}{Print helpful messages(default FALSE).}

\item{Y}{matrix of counts}
}
\value{
A object of class newModel
}
\description{
It creates different shared object and epxort them to the father
and the child process.
}
\keyword{internal}
